Use pre-existing substitution function
Switch the regex implementation used to be able to use the pcre-utils
version of the substitute function insteade of rolling our own. There
was a bug in the custom substitution method, essentially ignoring `$` in
regexes.

In order to maintain regex compatibility with the ones given from rails,
I still went with a PCRE based regex library, but the Regex type from
`pcre-light` wasn't the same as in `regex-pcre-builtin` so I had to swap
that around as well. `regex-pcre-builtin` exposes all the C-bindings as
IO, thus I had to introduce `unsafePerformIO` in this library as well.

Added tests for my failing cases and all tests now pass.